+++ src/dockerblade/files.py
@@ -142,6 +142,31 @@ class FileSystem:
 
             raise exc.UnexpectedError('failed to remove', error)
 
+    def write(self, filename: str, contents: Union[str, bytes]) -> None:
+        """Writes to a given file.
+
+        Parameters
+        ----------
+        filename: str
+            absolute path to the file.
+        contents: Union[str, bytes]
+            the text or binary contents of the file.
+        """
+        mode = 'wb' if isinstance(contents, bytes) else 'w'
+        directory = os.path.dirname(filename)
+        if not self.isdir(directory):
+            raise exc.ContainerFileNotFound(path=directory,
+                                            container_id=self.container.id)
+
+        # write to a temporary file on the host and copy to container
+        _, temp_filename = tempfile.mkstemp()
+        try:
+            with open(temp_filename, mode) as fh:
+                fh.write(contents)
+            self.copy_from_host(temp_filename, filename)
+        finally:
+            os.remove(temp_filename)
+
     @overload
     def read(self, filename: str) -> str:
         ...